[gc] mem: add mem cache alloced iterator
The GC needs to be able to iterate over all currently allocated objects.
When mem_cache is used, it stores all allocated objects and can be used
for this purpose.
Another mechanism possible for this is the memory object registry, but
as mem_cache improves performance, and is less memory intensive
than object registry, it is favorable.
Another approach would be to use platform specific knowledge to scan
the entire memory content, but in order to keep the gc simple, I would
like it to free js objects -- which would release their own related structures.