[gc] js: mark function's scope on mark phase
When walking the object tree, an enclosed function may have an
only reference to its scope (as the enclosing function had exited).
It should not be gc'd until the function is lost.
Therefore, we must walk the function scope.

Avoid a deadlock by making the walk cb return a true value if
an object had already been walked.