[build] support multiple IMAGE outputs
Some platforms require multiple image files, created by a single recipe.
It is useful to be able to specify those in the $(IMAGE) variable (though
probably not best named for this purpose).

Use a little trick with an intermediate non-existent target in order to make
sure the rule only runs once, allowing parallel build.