[esp8266] build: allow overriding default 'target link' recipe
The ESP8266 linking process is a bit strange -
The linker script is provided by the SDK and is quite complex, and
therefore I would like to avoid using a tailored linker script, but...
the .text section needs to be placed it at a specific section in the
flash (which is not pointed at by '.text' in the SDK linker script)
Therefore a little trickery is needed -- need to rename the .text
section after partially linking the code.

This obviously doesn't fit the 'default' linking process, therefore
providing a way to override it.